CodeQL | Fix information loss?

+++ game/core/src/tk/minersonline/history_survival/voxels/VoxelWorld.java
@@ -52,7 +52,7 @@ public class VoxelWorld implements RenderableProvider {
 		int len = CHUNK_SIZE_X * CHUNK_SIZE_Y * CHUNK_SIZE_Z * 6 * 6 / 3;
 		short[] indices = new short[len];
 		short j = 0;
-		for (i = 0; i < len; i += 6, j += 4) {
+		for (i = 0; i < len; i += 6, j += (short)4) {
 			indices[i + 0] = (short)(j + 0);
 			indices[i + 1] = (short)(j + 1);
 			indices[i + 2] = (short)(j + 2);
